+++ core/src/main/java/react4j/arez/ReactArezComponent.java
@@ -1,10 +1,7 @@
 package react4j.arez;
 
-import arez.Observer;
 import arez.annotations.Action;
-import arez.annotations.ObserverRef;
 import arez.annotations.PreDispose;
-import javax.annotation.Nonnull;
 import react4j.Component;
 
 /**
@@ -54,15 +51,6 @@ public abstract class ReactArezComponent
   {
     _renderDepsChanged = false;
   }
-
-  /**
-   * Return the Observer associated with the render tracker method.
-   *
-   * @return the Observer associated with the render tracker method.
-   */
-  @ObserverRef
-  @Nonnull
-  protected abstract Observer getRenderObserver();
 
   @PreDispose
   protected void preDispose()