+++ processor/src/test/resources/expected/com/example/arez/React4j_AllowNoArezDepsComponent.java
@@ -2,11 +2,13 @@ package com.example.arez;
 
 import arez.Arez;
 import arez.Disposable;
+import arez.Observer;
 import arez.annotations.ArezComponent;
 import arez.annotations.DepType;
 import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.Observe;
+import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
@@ -80,6 +82,10 @@ abstract class React4j_AllowNoArezDepsComponent extends AllowNoArezDepsComponent
     assert Disposable.isNotDisposed( this );
     return super.render();
   }
+
+  @Nonnull
+  @ObserverRef
+  abstract Observer getRenderObserver();
 
   @Override
   protected final void populateDebugData(@Nonnull final JsPropertyMap<Object> data) {