+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithProp.java
@@ -2,10 +2,12 @@ package com.example.arez;
 
 import arez.Arez;
 import arez.Disposable;
+import arez.Observer;
 import arez.annotations.ArezComponent;
 import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.Observe;
+import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
@@ -92,6 +94,10 @@ abstract class React4j_ComponentWithProp extends ComponentWithProp {
     }
     return result;
   }
+
+  @Nonnull
+  @ObserverRef
+  abstract Observer getRenderObserver();
 
   @Override
   protected final void populateDebugData(@Nonnull final JsPropertyMap<Object> data) {