+++ processor/src/test/resources/expected/com/example/arez/memoize/React4j_AlreadyPrioritizedMemoizeComponent.java
@@ -2,10 +2,12 @@ package com.example.arez.memoize;
 
 import arez.Arez;
 import arez.Disposable;
+import arez.Observer;
 import arez.annotations.ArezComponent;
 import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.Observe;
+import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
@@ -83,6 +85,10 @@ abstract class React4j_AlreadyPrioritizedMemoizeComponent extends AlreadyPriorit
     }
     return result;
   }
+
+  @Nonnull
+  @ObserverRef
+  abstract Observer getRenderObserver();
 
   @Override
   protected final void populateDebugData(@Nonnull final JsPropertyMap<Object> data) {