+++ processor/src/test/resources/expected/com/example/prop/React4j_NotObservableAsNotUpdateOnChangeProp.java
@@ -2,11 +2,13 @@ package com.example.prop;
 
 import arez.Arez;
 import arez.Disposable;
+import arez.Observer;
 import arez.annotations.ArezComponent;
 import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.Memoize;
 import arez.annotations.Observe;
+import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
@@ -93,6 +95,10 @@ abstract class React4j_NotObservableAsNotUpdateOnChangeProp extends NotObservabl
     }
     return result;
   }
+
+  @Nonnull
+  @ObserverRef
+  abstract Observer getRenderObserver();
 
   @Override
   protected final void populateDebugData(@Nonnull final JsPropertyMap<Object> data) {