+++ processor/src/test/resources/expected/com/example/prop/React4j_ObservableViaMemoizeProp.java
@@ -3,6 +3,7 @@ package com.example.prop;
 import arez.Arez;
 import arez.Disposable;
 import arez.ObservableValue;
+import arez.Observer;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.Executor;
@@ -11,6 +12,7 @@ import arez.annotations.Memoize;
 import arez.annotations.Observable;
 import arez.annotations.ObservableValueRef;
 import arez.annotations.Observe;
+import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
@@ -121,6 +123,10 @@ abstract class React4j_ObservableViaMemoizeProp extends ObservableViaMemoizeProp
     }
     return result;
   }
+
+  @Nonnull
+  @ObserverRef
+  abstract Observer getRenderObserver();
 
   @Override
   protected final void populateDebugData(@Nonnull final JsPropertyMap<Object> data) {