+++ processor/src/test/resources/expected/com/example/prop/React4j_ObservableViaObservedProp.java
@@ -3,6 +3,7 @@ package com.example.prop;
 import arez.Arez;
 import arez.Disposable;
 import arez.ObservableValue;
+import arez.Observer;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.Executor;
@@ -10,6 +11,7 @@ import arez.annotations.Feature;
 import arez.annotations.Observable;
 import arez.annotations.ObservableValueRef;
 import arez.annotations.Observe;
+import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
@@ -120,6 +122,10 @@ abstract class React4j_ObservableViaObservedProp extends ObservableViaObservedPr
     }
     return result;
   }
+
+  @Nonnull
+  @ObserverRef
+  abstract Observer getRenderObserver();
 
   @Override
   protected final void populateDebugData(@Nonnull final JsPropertyMap<Object> data) {